refactor: separate state from view model
