+++ presentation/src/main/java/com/madrid/presentation/viewModel/libraryViewModel/addtolist/MovieListUiState.kt
@@ -1,0 +1,28 @@
+package com.madrid.presentation.viewModel.libraryViewModel.addtolist
+
+import com.madrid.domain.entity.WatchList
+
+
+data class MovieListUiState(
+    val isLoading: Boolean = false,
+    val errorMessage: String? = null,
+    val successMessage: String? = null,
+    val createListSuccess: Boolean = false,
+    val addToListSuccess: Boolean = false,
+    val userLists: List<WatchList> = emptyList(),
+    val watchListItems: List<WatchListItemUiState> = emptyList(),
+    val isLoadingLists: Boolean = false
+)
+
+data class WatchListItemUiState(
+    val id: Int = 0,
+    val videosSize: Int = 0,
+    val watchListTitle: String = ""
+)
+
+sealed class MovieListEvent {
+    object ClearMessages : MovieListEvent()
+    object DismissNotification : MovieListEvent()
+    object LoadUserLists : MovieListEvent()
+}
+
